Reverted the initial value for the for loop back to 1 and changed numDigits to measure the number of digits in the square
The problem was not what I had initially thought.  The problem was that the loop needs to run for the length of the square because that is the number that is being changed, not the value inputed.  The change to numDigits fixed our problem, but with i = 0, the code would run one too many times. Changing it back to 1 fixed this problem.